python-api/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Bitcoin and cryptography (use same version as btc-vaultero)
bitcoin-utils @ git+https://github.com/karask/python-bitcoin-utils.git@b689ba7e8ef831803695b591a76cbfbf88504bd8

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# HTTP client for external API calls
httpx==0.25.2
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
